_oldarchive/old3/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# YouTube API & Parsing
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
yt-dlp==2023.12.30
beautifulsoup4==4.12.2
lxml==5.0.0

# Task Queue
celery[redis]==5.3.4
redis==4.6.0  # Совместимая версия с Celery 5.3.4
flower==2.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1

# NLP & AI
openai==1.6.1
transformers==4.36.2
# torch==2.1.2  # Закомментировано из-за конфликта версий

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
eventlet==0.36.1  # Для Celery на Windows

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Export formats
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.8